talha_chatbot_rag/pyproject.toml
[project]
name = "talha-chatbot-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.8",
    "groq>=0.9.0",
    "langchain>=0.2",
    "langchain-community>=0.2",
    "langchain-groq>=0.1.5",
    "langchain-text-splitters>=0.2",
    "pypdf>=4.2",
    "python-dotenv>=1.0",
    "sentence-transformers>=3.0",
    "streamlit>=1.36",
]
